requirements.txt
fastapi
uvicorn
SQLAlchemy
alembic
psycopg2-binary
pydantic
python-dotenv
Faker
typer
redis
tenacity
httpx
